Allow configuration of app.config at docker run time by putting a app.config file in the /home/dev/git/sirius-map-server/docker/openrouteservice/conf dir and mounting the file with '-v /home/dev/git/sirius-map-server/docker/openrouteservice/conf:/ors-conf'

+++ docker-entrypoint.sh
@@ -23,7 +23,9 @@ fi
 if [ -d "/usr/local/tomcat/webapps/ors" ]; then
 	cp -f /ors-conf/app.config $tomcat_appconfig
 else
-	cp -f $source_appconfig /ors-conf/app.config
+	if [ ! -f /ors-conf/app.config ]; then
+		cp -f $source_appconfig /ors-conf/app.config
+	fi
 	echo "### Package openrouteservice and deploy to Tomcat ###"
 	mvn -q -f /ors-core/openrouteservice/pom.xml package -DskipTests && \
 	cp -f /ors-core/openrouteservice/target/*.war /usr/local/tomcat/webapps/ors.war
